Changed the score to Double to admit decimal numbers

+++ app/src/main/java/com/mmag/bgamescoreboard/data/db/model/entities/Score.kt
@@ -15,5 +15,5 @@ data class Score(
     @ColumnInfo("category_id")
     val categoryId: Int,
     @ColumnInfo("score_amount")
-    val scoreAmount: Int
+    val scoreAmount: Double
 )
